requirements.txt
streamlit>=1.30.0
openai>=1.10.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
scipy>=1.10.0
scikit-learn>=1.3.0
lifelines>=0.27.0
matplotlib-venn>=0.11.0
adjustText>=0.8
python-dotenv>=1.0.0
openpyxl>=3.1.0
